Noon (#8)
* relax base upper bound

Tested working up to base-4.10 (ghc 8.2.2).

I've chosen to relax the bound to allow any base until the next major
version. Minor versions of base seem very unlikely to break this
library, since the only parts of base it depends on are the Prelude and
Data.Fixed. This is quite a common choice throughout hackage;
see https://packdeps.haskellers.com/reverse/base

closes https://github.com/intractable/horizon/issues/2

* add solarNoon

This is within a couple of minutes of values from
https://www.esrl.noaa.gov/gmd/grad/solcalc/ at the same longitude.
This module's sunrise and sunset calculations are likewise within a
couple of minutes to values from that site.

closes https://github.com/intractable/horizon/issues/3